feat(StyledLabel): loading and info variants added, (fetchApi): useFetchData.ts extracted
